refactor(map): extract 4 in-file sub-components to sub/ (2/7)
Step 2 of the PulseMapView strangler decomposition. Pure move — no
behaviour change. Component bodies and render output are byte-identical
to the prior inline definitions.

sub/AiStrip.tsx
  - The rose-tinted band: 6 render branches (underway / reorder / ready
    proposal / fetching / paused / single-stop) gated by aiState + lens
    + markerCount. Carries its own formatArrivalTime helper.
  - Internal state: whyExpanded, dragFromIdx, dragOverIdx, reorderListRef.

sub/LensEmptyState.tsx
  - Atlas-empty onboarding card (three paths), Atlas mid-geocoding wait
    card, Today/Week empty + "Switch to Atlas" CTA.

sub/LiveBroadcastSheet.tsx
  - Slide-up bottom sheet wrapping LiveTeamView. Wires useDialogA11y
    against sheetRef + closeBtnRef.

sub/ContactLocationPickerOverlay.tsx
  - Search-filtered contact picker dialog used by the Atlas empty-state
    "Pin a contact's home or work" path. Wires useDialogA11y against
    containerRef + searchInputRef.

PulseMapView.tsx
  - All 4 inline definitions removed (-786 lines).
  - Imports the 4 components from ./sub/* and continues to use them in
    the same render positions.
  - Lucide-react import shrinks from 16 icons to 3 (AlertTriangle,
    MapPinned, Users) — the other 13 lived only inside the extracted
    sub-components.
  - LiveTeamView + useDialogA11y imports drop — both now imported by
    the extracted sub-components directly.
  - AiProposal type drops from the host's aiTypes import — it was only
    used inside AiStrip's prop type, which now lives with the component.
  - End-of-file re-export `export { AiStrip, LiveBroadcastSheet }`
    preserves the e2e harness import surface; the harness keeps
    `import { AiStrip, LiveBroadcastSheet } from '../PulseMapView'`.
