+++ src/components/map/PulseMapView.tsx
@@ -1,23 +1,6 @@
 import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import { Circle, GoogleMap, Polygon, Polyline, useJsApiLoader } from '@react-google-maps/api';
-import {
-  AlertTriangle,
-  ArrowUpDown,
-  ChevronDown,
-  ChevronRight,
-  ChevronUp,
-  Globe,
-  Home,
-  MapPin,
-  MapPinned,
-  Navigation,
-  Radio,
-  Sparkles,
-  Sun,
-  Upload,
-  Users,
-  X,
-} from 'lucide-react';
+import { AlertTriangle, MapPinned, Users } from 'lucide-react';
 import { AppView, CalendarEvent, Contact, Thread } from '../../types';
 import { ContactCircle } from '../../types/contactCircleTypes';
 import { GOOGLE_MAPS_LIBRARIES, convexHull, getMapOptions, computeBounds } from '../../services/mapService';
@@ -46,8 +29,6 @@ import MapCircleOverlay from './contacts/MapCircleOverlay';
 import MapMeetingMarker from './contacts/MapMeetingMarker';
 import LocationEditModal from './contacts/LocationEditModal';
 import ImAtFAB from './sub/ImAtFAB';
-import LiveTeamView from './sub/LiveTeamView';
-import { useDialogA11y } from './sub/useDialogA11y';
 import {
   DAY_MS,
   DEFAULT_CENTER,
@@ -59,7 +40,11 @@ import {
   type MapLens,
   type MapViewMode,
 } from './sub/mapLens';
-import type { AcceptedRoute, AiProposal, AiState } from './sub/aiTypes';
+import type { AcceptedRoute, AiState } from './sub/aiTypes';
+import { AiStrip } from './sub/AiStrip';
+import { LensEmptyState } from './sub/LensEmptyState';
+import { LiveBroadcastSheet } from './sub/LiveBroadcastSheet';
+import { ContactLocationPickerOverlay } from './sub/ContactLocationPickerOverlay';
 
 interface PulseMapViewProps {
   contacts: Contact[];
@@ -1337,778 +1322,6 @@ const PulseMapView: React.FC<PulseMapViewProps> = ({
             onContactAction={onContactAction}
           />
         )}
-      </div>
-    </div>
-  );
-};
-
-// ============================================================
-// AI strip — Phase 2 wired. Renders one of:
-//   • Underway state (accepted route in flight)
-//   • Ready proposal (Gemini returned)
-//   • Quiet placeholder (TODAY + 2+ stops while AI is fetching)
-//   • Nothing (everything else)
-// Speed is craft — no spinners, no "AI is thinking…" theatre.
-// ============================================================
-
-interface AiStripProps {
-  lens: MapLens;
-  markerCount: number;
-  aiState: AiState;
-  acceptedRoute: AcceptedRoute | null;
-  acceptingRoute: boolean;
-  isDarkMode: boolean;
-  /** Available stops keyed by markerKey — used by the reorder list to render
-   *  human-readable labels without coupling AiStrip to MarkerData. */
-  stops: Array<{ id: string; label: string }>;
-  onAccept: () => void;
-  onDismissRoute: () => void;
-  onOpenInSystemMaps: () => void;
-  onReorderStart: () => void;
-  onReorderChange: (orderedIds: string[]) => void;
-  onReorderCancel: () => void;
-}
-
-function formatArrivalTime(d: Date): string {
-  return d.toLocaleTimeString(undefined, { hour: '2-digit', minute: '2-digit', hour12: false });
-}
-
-const AiStrip: React.FC<AiStripProps> = ({
-  lens,
-  markerCount,
-  aiState,
-  acceptedRoute,
-  acceptingRoute,
-  isDarkMode,
-  stops,
-  onAccept,
-  onDismissRoute,
-  onOpenInSystemMaps,
-  onReorderStart,
-  onReorderChange,
-  onReorderCancel,
-}) => {
-  // Local UI state — collapsed by default. Rationale expansion is per-mount
-  // (reset whenever the proposal swaps) so a fresh proposal doesn't surprise
-  // the user with someone else's reasoning still open.
-  const [whyExpanded, setWhyExpanded] = useState(false);
-  const [dragFromIdx, setDragFromIdx] = useState<number | null>(null);
-  // Cell over which the dragged item is currently hovering — drives the drop
-  // indicator without depending on browser-specific DragEvent.dataTransfer
-  // behaviour. Null when nothing is being hovered.
-  const [dragOverIdx, setDragOverIdx] = useState<number | null>(null);
-  // Ref to the reorder <ul> — keyboard moves use it to re-focus the row at
-  // the new index after React commits, so consecutive ArrowDown presses
-  // keep moving the SAME stop instead of whoever ends up at that position.
-  const reorderListRef = useRef<HTMLUListElement>(null);
-
-  const stripCls = `flex items-center gap-3 px-3 py-2 ${
-    isDarkMode
-      ? 'bg-gradient-to-r from-rose-500/10 via-rose-500/5 to-transparent'
-      : 'bg-gradient-to-r from-rose-50 via-rose-50/40 to-transparent'
-  }`;
-  const wrapperBorderCls = isDarkMode ? 'border-b border-rose-500/15' : 'border-b border-rose-500/20';
-  const monoStyle = { fontFamily: "'JetBrains Mono', monospace" } as const;
-
-  // Reset the Why? expansion when the underlying proposal changes (a new
-  // route arrives or we leave reorder mode). Otherwise the toggle would
-  // attach to whichever rationale happens to be in scope next.
-  const proposalSignature =
-    aiState.status === 'ready' && aiState.data.kind === 'route'
-      ? aiState.data.proposal.summary
-      : '';
-  useEffect(() => { setWhyExpanded(false); }, [proposalSignature]);
-
-  // 1. Underway — accepted route in flight, takes precedence over any pending
-  //    AI fetch. Two buttons: Open in Maps + Dismiss.
-  if (acceptedRoute) {
-    const arriving = formatArrivalTime(acceptedRoute.arrivesAt);
-    return (
-      <div className={`${stripCls} ${wrapperBorderCls}`} role="status" aria-live="polite">
-        <Navigation size={14} className="text-rose-500 flex-shrink-0" />
-        <span
-          className="text-[10px] tracking-[0.1em] uppercase text-rose-500 flex-shrink-0"
-          style={monoStyle}
-        >
-          PULSE AI · UNDERWAY
-        </span>
-        <span className={`text-xs flex-1 truncate ${isDarkMode ? 'text-gray-200' : 'text-gray-800'}`}>
-          {acceptedRoute.orderedMarkerKeys.length} stops · {acceptedRoute.durationMin} min · arriving {arriving}
-        </span>
-        <div className="flex items-center gap-1 flex-shrink-0">
-          <button
-            type="button"
-            onClick={onOpenInSystemMaps}
-            className="inline-flex items-center gap-1 px-2.5 py-1 rounded text-[11px] font-semibold bg-rose-500 text-white hover:bg-rose-600 transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-rose-500 focus-visible:ring-offset-1"
-          >
-            Open in Maps
-            <ChevronRight size={11} />
-          </button>
-          <button
-            type="button"
-            onClick={onDismissRoute}
-            aria-label="Dismiss route"
-            className={`p-1.5 rounded transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-rose-500 focus-visible:ring-offset-1 ${
-              isDarkMode ? 'hover:bg-white/10 text-gray-400' : 'hover:bg-rose-500/10 text-gray-500'
-            }`}
-          >
-            <X size={12} aria-hidden="true" />
-          </button>
-        </div>
-      </div>
-    );
-  }
-
-  // 2. Reorder mode — drag-to-reorder OR keyboard ArrowUp/Down via per-row
-  //    buttons. Drop / keyboard / button paths all route through moveStop so
-  //    behaviour stays one shape. Accept replays DirectionsService against
-  //    the new sequence; Cancel reverts. ────────────────────────────────────
-  if (aiState.status === 'reordering') {
-    const moveStop = (fromIdx: number, toIdx: number) => {
-      if (fromIdx === toIdx) return;
-      if (toIdx < 0 || toIdx >= aiState.orderedIds.length) return;
-      const next = aiState.orderedIds.slice();
-      const [moved] = next.splice(fromIdx, 1);
-      next.splice(toIdx, 0, moved);
-      onReorderChange(next);
-    };
-    const handleDrop = (toIdx: number) => {
-      if (dragFromIdx == null) {
-        setDragOverIdx(null);
-        return;
-      }
-      moveStop(dragFromIdx, toIdx);
-      setDragFromIdx(null);
-      setDragOverIdx(null);
-    };
-    // Re-focus the same row after a keyboard move so consecutive ArrowDown
-    // presses continue moving the SAME stop, not the one now at that index.
-    const focusRowByIdx = (listEl: HTMLUListElement | null, idx: number) => {
-      if (!listEl) return;
-      const rows = listEl.querySelectorAll<HTMLLIElement>('li[data-reorder-row="1"]');
-      rows[idx]?.focus();
-    };
-    const validCount = aiState.orderedIds.filter(id => stops.some(s => s.id === id)).length;
-    return (
-      <div className={wrapperBorderCls}>
-        <div className={stripCls} role="status" aria-live="polite">
-          <ArrowUpDown size={14} className="text-rose-500 flex-shrink-0" aria-hidden="true" />
-          <span
-            className="text-[10px] tracking-[0.1em] uppercase text-rose-500 flex-shrink-0"
-            style={monoStyle}
-          >
-            PULSE AI · REORDER
-          </span>
-          <span className={`text-xs flex-1 truncate ${isDarkMode ? 'text-gray-300' : 'text-gray-700'}`}>
-            Drag or use arrow keys, then Accept.
-          </span>
-          <div className="flex items-center gap-1 flex-shrink-0">
-            <button
-              type="button"
-              onClick={onReorderCancel}
-              className={`px-2.5 py-1 rounded text-[11px] font-medium transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-rose-500 focus-visible:ring-offset-1 ${
-                isDarkMode ? 'text-gray-300 hover:bg-white/10' : 'text-gray-600 hover:bg-rose-500/10'
-              }`}
-            >
-              Cancel
-            </button>
-            <button
-              type="button"
-              onClick={onAccept}
-              disabled={acceptingRoute || validCount < 2}
-              className="px-2.5 py-1 rounded text-[11px] font-semibold bg-rose-500 text-white hover:bg-rose-600 transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-rose-500 focus-visible:ring-offset-1 disabled:opacity-60 disabled:cursor-not-allowed"
-            >
-              {acceptingRoute ? 'Routing…' : 'Accept'}
-            </button>
-          </div>
-        </div>
-        <p id="reorder-instructions" className="sr-only">
-          Use the up and down buttons or press ArrowUp / ArrowDown while a row
-          is focused to reorder. Home jumps to the first stop, End to the last.
-          Then press Accept to apply.
-        </p>
-        <ul
-          className={`px-3 pb-2 pt-1 space-y-1 ${
-            isDarkMode ? 'bg-rose-500/[0.04]' : 'bg-rose-50/40'
-          }`}
-          aria-label="Reorder stops"
-          aria-describedby="reorder-instructions"
-          ref={reorderListRef}
-        >
-          {aiState.orderedIds.map((id, idx) => {
-            const stop = stops.find(s => s.id === id);
-            if (!stop) return null;
-            const isDragging = dragFromIdx === idx;
-            const isOver = dragOverIdx === idx && dragFromIdx !== idx;
-            const total = aiState.orderedIds.length;
-            const atTop = idx === 0;
-            const atBottom = idx === total - 1;
-            const handleKey = (e: React.KeyboardEvent<HTMLLIElement>) => {
-              // Ignore key events bubbling from the Up/Down buttons — they
-              // handle their own click via Enter/Space natively.
-              if ((e.target as HTMLElement).tagName === 'BUTTON') return;
-              let target: number | null = null;
-              if (e.key === 'ArrowUp' && !atTop) target = idx - 1;
-              else if (e.key === 'ArrowDown' && !atBottom) target = idx + 1;
-              else if (e.key === 'Home' && !atTop) target = 0;
-              else if (e.key === 'End' && !atBottom) target = total - 1;
-              if (target == null) return;
-              e.preventDefault();
-              moveStop(idx, target);
-              // Defer focus to after React commits the new order.
-              requestAnimationFrame(() => focusRowByIdx(reorderListRef.current, target));
-            };
-            return (
-              <li
-                key={id}
-                data-reorder-row="1"
-                tabIndex={0}
-                role="listitem"
-                aria-label={`Stop ${idx + 1} of ${total}: ${stop.label}`}
-                onKeyDown={handleKey}
-                draggable
-                onDragStart={(e) => {
-                  setDragFromIdx(idx);
-                  e.dataTransfer.effectAllowed = 'move';
-                  // Some browsers need a payload to initiate the drag image.
-                  try { e.dataTransfer.setData('text/plain', id); } catch { /* IE-era safety */ }
-                }}
-                onDragOver={(e) => {
-                  e.preventDefault();
-                  e.dataTransfer.dropEffect = 'move';
-                  setDragOverIdx(idx);
-                }}
-                onDragLeave={() => {
-                  setDragOverIdx(prev => (prev === idx ? null : prev));
-                }}
-                onDrop={(e) => {
-                  e.preventDefault();
-                  handleDrop(idx);
-                }}
-                onDragEnd={() => {
-                  setDragFromIdx(null);
-                  setDragOverIdx(null);
-                }}
-                className={`flex items-center gap-2.5 px-2 py-1.5 rounded cursor-grab active:cursor-grabbing select-none transition-all focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-rose-500 focus-visible:ring-offset-1 ${
-                  isDarkMode
-                    ? 'bg-zinc-900/60 border border-white/5 hover:border-rose-500/30'
-                    : 'bg-white border border-rose-100 hover:border-rose-300'
-                } ${isDragging ? 'opacity-40' : ''} ${isOver ? 'ring-2 ring-rose-500/60 ring-offset-0' : ''}`}
-              >
-                <span
-                  className={`inline-flex items-center justify-center w-5 h-5 rounded-full text-[10px] font-semibold text-rose-500 ${
-                    isDarkMode ? 'bg-rose-500/15' : 'bg-rose-100'
-                  }`}
-                  style={monoStyle}
-                  aria-hidden="true"
-                >
-                  {idx + 1}
-                </span>
-                <span
-                  className={`text-xs flex-1 truncate ${isDarkMode ? 'text-gray-200' : 'text-gray-800'}`}
-                  aria-hidden="true"
-                >
-                  {stop.label}
-                </span>
-                <button
-                  type="button"
-                  onClick={() => {
-                    moveStop(idx, idx - 1);
-                    requestAnimationFrame(() => focusRowByIdx(reorderListRef.current, idx - 1));
-                  }}
-                  disabled={atTop}
-                  aria-label={`Move ${stop.label} up`}
-                  className={`p-1 rounded transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-rose-500 focus-visible:ring-offset-1 disabled:opacity-30 disabled:cursor-not-allowed ${
-                    isDarkMode ? 'text-gray-400 hover:bg-white/10 hover:text-rose-300' : 'text-gray-500 hover:bg-rose-500/10 hover:text-rose-600'
-                  }`}
-                >
-                  <ChevronUp size={12} aria-hidden="true" />
-                </button>
-                <button
-                  type="button"
-                  onClick={() => {
-                    moveStop(idx, idx + 1);
-                    requestAnimationFrame(() => focusRowByIdx(reorderListRef.current, idx + 1));
-                  }}
-                  disabled={atBottom}
-                  aria-label={`Move ${stop.label} down`}
-                  className={`p-1 rounded transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-rose-500 focus-visible:ring-offset-1 disabled:opacity-30 disabled:cursor-not-allowed ${
-                    isDarkMode ? 'text-gray-400 hover:bg-white/10 hover:text-rose-300' : 'text-gray-500 hover:bg-rose-500/10 hover:text-rose-600'
-                  }`}
-                >
-                  <ChevronDown size={12} aria-hidden="true" />
-                </button>
-              </li>
-            );
-          })}
-        </ul>
-      </div>
-    );
-  }
-
-  // 3. Ready proposal.
-  if (aiState.status === 'ready') {
-    const data = aiState.data;
-    if (data.kind === 'route') {
-      const summary = data.proposal.summary;
-      const rationale = data.proposal.rationale;
-      const count = data.proposal.orderedIds.length;
-      return (
-        <div className={wrapperBorderCls}>
-          <div className={stripCls} role="status" aria-live="polite">
-            <Sparkles size={14} className="text-rose-500 flex-shrink-0" />
-            <span className="text-[10px] tracking-[0.1em] uppercase text-rose-500 flex-shrink-0" style={monoStyle}>
-              PULSE AI · ROUTE
-            </span>
-            <span
-              className={`text-xs flex-1 truncate ${isDarkMode ? 'text-gray-200' : 'text-gray-800'}`}
-              title={rationale ?? summary}
-            >
-              {summary}
-            </span>
-            <div className="flex items-center gap-1 flex-shrink-0">
-              {rationale && (
-                <button
-                  type="button"
-                  onClick={() => setWhyExpanded(v => !v)}
-                  aria-expanded={whyExpanded}
-                  aria-controls="pulse-ai-rationale"
-                  className={`inline-flex items-center gap-0.5 px-2 py-1.5 rounded text-[11px] font-medium transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-rose-500 focus-visible:ring-offset-1 ${
-                    isDarkMode ? 'text-gray-400 hover:text-rose-300 hover:bg-white/5' : 'text-gray-500 hover:text-rose-600 hover:bg-rose-500/5'
-                  }`}
-                >
-                  Why?
-                  <ChevronDown
-                    size={11}
-                    aria-hidden="true"
-                    className={`transition-transform ${whyExpanded ? 'rotate-180' : ''}`}
-                  />
-                </button>
-              )}
-              <button
-                type="button"
-                onClick={onReorderStart}
-                disabled={acceptingRoute || count < 2}
-                className={`inline-flex items-center gap-1 px-2 py-1 rounded text-[11px] font-medium transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-rose-500 focus-visible:ring-offset-1 disabled:opacity-50 disabled:cursor-not-allowed ${
-                  isDarkMode ? 'text-gray-300 hover:bg-white/10' : 'text-gray-600 hover:bg-rose-500/10'
-                }`}
-              >
-                <ArrowUpDown size={11} />
-                Reorder
-              </button>
-              <button
-                type="button"
-                onClick={onAccept}
-                disabled={acceptingRoute || count < 2}
-                className="px-2.5 py-1 rounded text-[11px] font-semibold bg-rose-500 text-white hover:bg-rose-600 transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-rose-500 focus-visible:ring-offset-1 disabled:opacity-60 disabled:cursor-not-allowed"
-              >
-                {acceptingRoute ? 'Routing…' : 'Accept'}
-              </button>
-            </div>
-          </div>
-          {rationale && (
-            <div
-              id="pulse-ai-rationale"
-              aria-hidden={!whyExpanded}
-              className={isDarkMode ? 'bg-rose-500/[0.04]' : 'bg-rose-50/40'}
-              style={{
-                maxHeight: whyExpanded ? 240 : 0,
-                overflow: 'hidden',
-                transition: 'max-height 220ms cubic-bezier(0.16, 1, 0.3, 1)',
-              }}
-            >
-              <p
-                className={`px-3 py-2 text-xs leading-relaxed ${
-                  isDarkMode ? 'text-gray-300' : 'text-gray-700'
-                }`}
-                style={{ maxWidth: '70ch' }}
-              >
-                {rationale}
-              </p>
-            </div>
-          )}
-        </div>
-      );
-    }
-    if (data.kind === 'plan' || data.kind === 'insight') {
-      const label = data.kind === 'plan' ? 'PULSE AI · PLAN' : 'PULSE AI · INSIGHT';
-      return (
-        <div className={`${stripCls} ${wrapperBorderCls}`} role="status" aria-live="polite">
-          <Sparkles size={14} className="text-rose-500 flex-shrink-0" />
-          <span className="text-[10px] tracking-[0.1em] uppercase text-rose-500 flex-shrink-0" style={monoStyle}>
-            {label}
-          </span>
-          <span className={`text-xs flex-1 truncate ${isDarkMode ? 'text-gray-200' : 'text-gray-800'}`}>
-            {data.proposal.summary}
-          </span>
-        </div>
-      );
-    }
-  }
-
-  // 4. Fetching placeholder — only for TODAY with >=2 stops, where a route
-  //    proposal is genuinely imminent. Other lenses fail silent.
-  if (aiState.status === 'fetching' && lens === 'today' && markerCount >= 2) {
-    return (
-      <div className={`${stripCls} ${wrapperBorderCls}`} role="status" aria-live="polite">
-        <Sparkles size={14} className={`flex-shrink-0 ${isDarkMode ? 'text-rose-500/70' : 'text-rose-500/80'}`} />
-        <span className="text-[10px] tracking-[0.1em] uppercase text-rose-500/70 flex-shrink-0" style={monoStyle}>
-          PULSE AI · ROUTE
-        </span>
-        <span className={`text-xs flex-1 truncate ${isDarkMode ? 'text-gray-400' : 'text-gray-500'}`}>
-          {markerCount} stops on the map.
-        </span>
-      </div>
-    );
-  }
-
-  // 5. Paused — workspace hit the AI cap. Tell the truth, don't bark.
-  if (aiState.status === 'paused' && lens === 'today' && markerCount >= 2) {
-    return (
-      <div className={`${stripCls} ${wrapperBorderCls}`} role="status" aria-live="polite">
-        <Sparkles size={14} className="text-rose-500/60 flex-shrink-0" />
-        <span className="text-[10px] tracking-[0.1em] uppercase text-rose-500/60 flex-shrink-0" style={monoStyle}>
-          PULSE AI · PAUSED
-        </span>
-        <span className={`text-xs flex-1 truncate ${isDarkMode ? 'text-gray-400' : 'text-gray-500'}`}>
-          Route proposals on a short break. Resuming after the workspace cap clears.
-        </span>
-      </div>
-    );
-  }
-
-  // 6. Today + 1 stop — declare it honestly even when AI returned none.
-  if (lens === 'today' && markerCount === 1) {
-    return (
-      <div className={`${stripCls} ${wrapperBorderCls}`} role="status" aria-live="polite">
-        <Sparkles size={14} className="text-rose-500/70 flex-shrink-0" />
-        <span className="text-[10px] tracking-[0.1em] uppercase text-rose-500/70 flex-shrink-0" style={monoStyle}>
-          PULSE AI · ROUTE
-        </span>
-        <span className={`text-xs flex-1 truncate ${isDarkMode ? 'text-gray-400' : 'text-gray-500'}`}>
-          No route today, just one stop.
-        </span>
-      </div>
-    );
-  }
-
-  // Everything else — hide.
-  return null;
-};
-
-// ============================================================
-// Lens-specific empty state. Replaces the prior single-CTA card
-// with copy + paths matched to the active lens.
-// ============================================================
-
-interface LensEmptyStateProps {
-  lens: MapLens;
-  isDarkMode: boolean;
-  atlasHasAnyPinned: boolean;
-  hasGeocodingInFlight: boolean;
-  canAddLocation: boolean;
-  onOpenAtlas: () => void;
-  onAutoGeocode: () => void;
-  onPinWhereIAm: () => void;
-  onPickContact: () => void;
-}
-
-const LensEmptyState: React.FC<LensEmptyStateProps> = ({
-  lens,
-  isDarkMode,
-  atlasHasAnyPinned,
-  hasGeocodingInFlight,
-  canAddLocation,
-  onOpenAtlas,
-  onAutoGeocode,
-  onPinWhereIAm,
-  onPickContact,
-}) => {
-  // Atlas-empty (first-run, no pins anywhere): three-path onboarding.
-  if (lens === 'atlas' && !atlasHasAnyPinned) {
-    return (
-      <div className="absolute inset-0 flex flex-col items-center justify-center pointer-events-none px-6">
-        <div
-          className={`rounded-2xl px-6 py-5 shadow-lg backdrop-blur-2xl border max-w-md pointer-events-auto ${
-            isDarkMode ? 'bg-zinc-950/85 border-white/10' : 'bg-white/90 border-gray-200'
-          }`}
-        >
-          <MapPin size={28} className="text-rose-500/70 mb-3" />
-          <p className={`text-sm font-semibold mb-1 ${isDarkMode ? 'text-gray-100' : 'text-gray-900'}`}>
-            Pin your network to see distance, routes, and circles.
-          </p>
-          <p className={`text-xs mb-4 ${isDarkMode ? 'text-gray-500' : 'text-gray-500'}`}>
-            Pick a starting move — the rest fills in as you go.
-          </p>
-          <div className="flex flex-col gap-2">
-            {hasGeocodingInFlight && (
-              <button
-                type="button"
-                onClick={onAutoGeocode}
-                className="flex items-center gap-2 px-3 py-2 rounded-lg text-xs font-semibold bg-rose-500 text-white hover:bg-rose-600 transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-rose-500 focus-visible:ring-offset-1"
-              >
-                <Sparkles size={12} />
-                Auto-geocode contacts with addresses
-              </button>
-            )}
-            <button
-              type="button"
-              onClick={onPinWhereIAm}
-              className={`flex items-center gap-2 px-3 py-2 rounded-lg text-xs font-medium border transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-rose-500 ${
-                isDarkMode ? 'border-white/10 text-gray-200 hover:border-rose-500/40 hover:text-rose-300' : 'border-gray-200 text-gray-700 hover:border-rose-500/40 hover:text-rose-600'
-              }`}
-            >
-              <Home size={12} />
-              Pin a contact's home or work
-            </button>
-            <button
-              type="button"
-              onClick={onPickContact}
-              disabled={!canAddLocation}
-              className={`flex items-center gap-2 px-3 py-2 rounded-lg text-xs font-medium border transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-rose-500 disabled:opacity-50 disabled:cursor-not-allowed ${
-                isDarkMode ? 'border-white/10 text-gray-200 hover:border-rose-500/40 hover:text-rose-300' : 'border-gray-200 text-gray-700 hover:border-rose-500/40 hover:text-rose-600'
-              }`}
-            >
-              <Upload size={12} />
-              Import addresses from a list
-            </button>
-          </div>
-        </div>
-      </div>
-    );
-  }
-
-  // Atlas mid-state: addresses queued for geocoding.
-  if (lens === 'atlas' && hasGeocodingInFlight) {
-    return (
-      <div className="absolute inset-0 flex items-center justify-center pointer-events-none">
-        <div
-          className={`rounded-2xl px-6 py-5 text-center shadow-lg backdrop-blur-2xl border max-w-sm pointer-events-auto ${
-            isDarkMode ? 'bg-zinc-950/85 border-white/10' : 'bg-white/90 border-gray-200'
-          }`}
-        >
-          <MapPinned size={28} className="text-rose-500/70 mx-auto mb-3 motion-safe:animate-pulse" />
-          <p className={`text-sm font-semibold ${isDarkMode ? 'text-gray-200' : 'text-gray-800'}`}>
-            Geocoding addresses…
-          </p>
-          <p className={`text-xs mt-1 ${isDarkMode ? 'text-gray-500' : 'text-gray-500'}`}>
-            Pins appear as Google resolves each contact.
-          </p>
-        </div>
-      </div>
-    );
-  }
-
-  // Today / Week empty when Atlas has pins: tell the truth, offer the lens swap.
-  return (
-    <div className="absolute inset-0 flex items-center justify-center pointer-events-none">
-      <div
-        className={`rounded-2xl px-6 py-5 text-center shadow-lg backdrop-blur-2xl border max-w-sm pointer-events-auto ${
-          isDarkMode ? 'bg-zinc-950/85 border-white/10' : 'bg-white/90 border-gray-200'
-        }`}
-      >
-        <Sun size={28} className="text-rose-500/70 mx-auto mb-3" />
-        <p className={`text-sm font-semibold mb-1 ${isDarkMode ? 'text-gray-100' : 'text-gray-900'}`}>
-          {lens === 'today' ? 'Nothing on the map today.' : 'Nothing on the map this week.'}
-        </p>
-        <p className={`text-xs mb-4 ${isDarkMode ? 'text-gray-500' : 'text-gray-500'}`}>
-          Open Atlas to browse your full network.
-        </p>
-        <button
-          type="button"
-          onClick={onOpenAtlas}
-          className={`inline-flex items-center gap-1.5 px-3 py-1.5 rounded-lg text-xs font-medium border transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-rose-500 ${
-            isDarkMode ? 'border-white/15 text-gray-200 hover:border-rose-500/40 hover:text-rose-300' : 'border-gray-200 text-gray-700 hover:border-rose-500/40 hover:text-rose-600'
-          }`}
-        >
-          <Globe size={11} />
-          Switch to Atlas
-        </button>
-      </div>
-    </div>
-  );
-};
-
-// ============================================================
-// Live broadcast sheet — slide-up wrapper around LiveTeamView.
-// Replaces the previous peer-lens treatment of live presence.
-// ============================================================
-
-interface LiveBroadcastSheetProps {
-  contacts: Contact[];
-  liveLocations: Map<string, UserLocation>;
-  isDarkMode: boolean;
-  onClose: () => void;
-  onContactAction: (action: 'message' | 'vox' | 'meet', contactId: string) => void;
-}
-
-const LiveBroadcastSheet: React.FC<LiveBroadcastSheetProps> = ({
-  contacts,
-  liveLocations,
-  isDarkMode,
-  onClose,
-  onContactAction,
-}) => {
-  const sheetRef = useRef<HTMLDivElement>(null);
-  const closeBtnRef = useRef<HTMLButtonElement>(null);
-  useDialogA11y({ containerRef: sheetRef, onClose, initialFocusRef: closeBtnRef });
-
-  return (
-    <div
-      className="absolute inset-0 z-30 flex items-end justify-center"
-      style={{ background: 'rgba(0,0,0,0.45)' }}
-      onClick={onClose}
-      role="dialog"
-      aria-modal="true"
-      aria-labelledby="live-broadcasters-title"
-    >
-      <div
-        ref={sheetRef}
-        className={`w-full max-w-2xl rounded-t-2xl border-t border-x shadow-2xl overflow-hidden map-sheet-up ${
-          isDarkMode ? 'bg-zinc-950 border-white/10' : 'bg-white border-gray-200'
-        }`}
-        style={{ maxHeight: '70%' }}
-        onClick={(e) => e.stopPropagation()}
-      >
-        <div className={`flex items-center justify-between px-4 py-3 border-b ${isDarkMode ? 'border-white/10' : 'border-gray-100'}`}>
-          <div className="flex items-center gap-2">
-            <Radio size={14} className="text-rose-500" aria-hidden="true" />
-            <span
-              id="live-broadcasters-title"
-              className="text-[11px] tracking-[0.1em] uppercase"
-              style={{ fontFamily: "'JetBrains Mono', monospace" }}
-            >
-              Broadcasting now
-            </span>
-          </div>
-          <button
-            ref={closeBtnRef}
-            type="button"
-            onClick={onClose}
-            aria-label="Close"
-            className={`p-1.5 rounded transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-rose-500 focus-visible:ring-offset-1 ${
-              isDarkMode ? 'hover:bg-white/10 text-gray-400' : 'hover:bg-gray-100 text-gray-500'
-            }`}
-          >
-            <X size={14} aria-hidden="true" />
-          </button>
-        </div>
-        <div className="h-[60vh] overflow-y-auto">
-          <LiveTeamView
-            contacts={contacts}
-            liveLocations={liveLocations}
-            isDarkMode={isDarkMode}
-            onContactAction={onContactAction}
-          />
-        </div>
-      </div>
-    </div>
-  );
-};
-
-// ============================================================
-// Contact picker overlay (empty-state path). Inline because it's
-// tightly coupled to the empty state and not reused elsewhere.
-// ============================================================
-
-interface ContactLocationPickerOverlayProps {
-  contacts: Contact[];
-  isDarkMode: boolean;
-  onClose: () => void;
-  onPick: (contactId: string) => void;
-}
-
-const ContactLocationPickerOverlay: React.FC<ContactLocationPickerOverlayProps> = ({
-  contacts,
-  isDarkMode,
-  onClose,
-  onPick,
-}) => {
-  const [query, setQuery] = useState('');
-  const filtered = contacts.filter(c =>
-    !query || c.name.toLowerCase().includes(query.toLowerCase())
-  );
-
-  const containerRef = useRef<HTMLDivElement>(null);
-  const searchInputRef = useRef<HTMLInputElement>(null);
-  // Shared dialog a11y — focus trap, Escape, restore focus to trigger on
-  // unmount. Initial focus goes to the search field so the operator can
-  // immediately filter the contact list.
-  useDialogA11y({ containerRef, onClose, initialFocusRef: searchInputRef });
-
-  return (
-    <div
-      className="absolute inset-0 z-20 flex items-center justify-center p-6"
-      style={{ background: 'rgba(0,0,0,0.45)' }}
-      onClick={onClose}
-      role="dialog"
-      aria-modal="true"
-      aria-labelledby="contact-picker-title"
-    >
-      <div
-        ref={containerRef}
-        className={`w-full max-w-md rounded-2xl shadow-2xl border overflow-hidden ${
-          isDarkMode ? 'bg-zinc-900 border-white/10' : 'bg-white border-gray-200'
-        }`}
-        onClick={e => e.stopPropagation()}
-      >
-        <div className={`px-5 py-4 border-b ${isDarkMode ? 'border-white/10' : 'border-gray-100'}`}>
-          <h3
-            id="contact-picker-title"
-            className={`text-base font-semibold ${isDarkMode ? 'text-white' : 'text-gray-900'}`}
-          >
-            Pick a contact to locate
-          </h3>
-          <input
-            ref={searchInputRef}
-            type="text"
-            value={query}
-            onChange={(e) => setQuery(e.target.value)}
-            placeholder="Search by name…"
-            aria-label="Search contacts by name"
-            className={`mt-3 w-full px-3 py-2 rounded-lg text-sm outline-none transition-colors ${
-              isDarkMode
-                ? 'bg-white/5 border border-white/10 text-white placeholder-gray-500 focus:border-rose-500'
-                : 'bg-gray-50 border border-gray-200 text-gray-900 placeholder-gray-400 focus:border-rose-500'
-            }`}
-          />
-        </div>
-        <div className="max-h-80 overflow-y-auto">
-          {filtered.length === 0 ? (
-            <p className={`px-5 py-8 text-sm text-center ${isDarkMode ? 'text-gray-500' : 'text-gray-400'}`}>
-              No contacts match your search.
-            </p>
-          ) : (
-            filtered.map(c => (
-              <button
-                key={c.id}
-                type="button"
-                onClick={() => onPick(c.id)}
-                className={`w-full flex items-center gap-3 px-5 py-2.5 text-left transition-colors focus-visible:outline-none focus-visible:bg-rose-500/10 focus-visible:ring-2 focus-visible:ring-inset focus-visible:ring-rose-500 ${
-                  isDarkMode ? 'hover:bg-white/5' : 'hover:bg-gray-50'
-                }`}
-              >
-                <div
-                  aria-hidden="true"
-                  className="w-8 h-8 rounded-full flex items-center justify-center text-white text-xs font-semibold flex-shrink-0"
-                  style={{ backgroundColor: c.avatarColor || '#f43f5e' }}
-                >
-                  {c.name.split(' ').map(n => n[0]).join('').slice(0, 2).toUpperCase()}
-                </div>
-                <div className="min-w-0">
-                  <p className={`text-sm font-medium truncate ${isDarkMode ? 'text-white' : 'text-gray-900'}`}>
-                    {c.name}
-                  </p>
-                  {c.role && (
-                    <p className={`text-xs truncate ${isDarkMode ? 'text-gray-500' : 'text-gray-500'}`}>
-                      {c.role}
-                    </p>
-                  )}
-                </div>
-              </button>
-            ))
-          )}
-        </div>
       </div>
     </div>
   );
